src/imitation_learning/requirements_additional.txt
# Training and Inference Dependencies for Tracer RC Car ACT Model
# Add these to your existing lerobot environment

# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0

# Data processing
pandas>=1.3.0
pillow>=8.0.0
opencv-python>=4.5.0

# Visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0

# For edge deployment and optimization
onnx>=1.12.0
onnxruntime>=1.12.0

# Optional: For quantization and optimization
torch-pruning
torchinfo

# Optional: For advanced monitoring
wandb
tensorboard